# Yum/src/main/resources/static/api/swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.2.7",
        "title": "Yum Food Orders",
        "license": {
            "name": "GPLv3",
            "url": "http://www.gnu.org/licenses/"
        },
        "description": "**Yum application, order food daily from the best chef in town**\n\nThis API is used by the angular.io client, and is not meant to be used otherwise.\n\nFind source code of this API [here](http://gitlab/)\n\nCopyright (C) 2017 JR Technologies.\n\n  \n------------------------------------\n\n    Last edit: 18/05/2017 13:30\n\n-------------------------------------\n"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer": {
            "description": "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header.\nA valid JWT token is generated by the API and returned as answer of a call to the path /api/auth/login giving a valid user & password.\nThe following syntax must be used in the 'Authorization' header :\nBearer xxxxxx.yyyyyyy.zzzzzz\n",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/settings": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "description": "get user settings",
                "responses": {
                    "200": {
                        "description": "User data",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "description": "set user data/settings",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User data changed successfully"
                    },
                    "400": {
                        "description": "Bad reset password data",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Concurrent modification error",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/settings/picture": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "summary": "Uploads profile pic.",
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "file",
                        "type": "file",
                        "description": "The file to upload"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Picture set"
                    },
                    "404": {
                        "description": "no picture"
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "summary": "deletes profile picture.",
                "responses": {
                    "204": {
                        "description": "Picture deleted"
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/settings/picture/token": {
            "get": {
                "tags": [
                    "hungry"
                ],
                "description": "Get profile picture.",
                "parameters": [
                    {
                        "name": "token",
                        "description": "auth token",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "image/jpeg"
                ],
                "responses": {
                    "200": {
                        "description": "A jpeg file.",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Invalid token"
                    },
                    "404": {
                        "description": "image not found"
                    }
                }
            }
        },
        "/menus/weekly": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "summary": "Gets weekly menus",
                "description": "Returns a list containing all menus of the week.",
                "responses": {
                    "200": {
                        "description": "A list of daily menus",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DailyMenu"
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/menus/weekly/{week}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "summary": "Gets weekly menu of the specified week",
                "description": "Returns a list containing all menus of the week.",
                "parameters": [
                    {
                        "name": "week",
                        "description": "ex. 02-2017",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of daily menus",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DailyMenu"
                            }
                        }
                    },
                    "400": {
                        "description": "Weekly menu not found."
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/menus/monthly": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "summary": "Gets monthly menus.",
                "description": "Return a list containing all menus of the month.",
                "responses": {
                    "200": {
                        "description": "A list of daily menus.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DailyMenu"
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
        "/menus/monthly/{monthyear}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "hungry"
                ],
                "summary": "Gets monthly menus of the specified month.",
                "description": "Returns a list containing all menus of the month.",
                "parameters": [
                    {
                        "name": "monthyear",
                        "description": "ex. 12-2016",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of daily menus.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DailyMenu"
                            }
                        }
                    },
                    "400": {
                        "description": "Monthly menu not found."
                    },
                    "500": {
                        "$ref": "#/responses/Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "description": "User DTO",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "lastName": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string",
                    "format": "date"
                },
                "approved": {
                    "type": "boolean"
                },
                "hasPicture": {
                    "type": "boolean"
                },
                "lastEdit": {
                    "$ref": "#/definitions/LastEdit"
                }
            }
        },
...
        "UserSettings": {
            "type": "object",
            "description": "User DTO",
            "properties": {
                "lastName": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "lastEdit": {
                    "$ref": "#/definitions/LastEdit"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "description": "An error with some info DTO",
            "properties": {
                "message": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "FoodWithQuantity": {
            "type": "object",
            "description": "Food with quantity",
            "properties": {
                "food": {
                    "$ref": "#/definitions/Food"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "Food": {
            "type": "object",
            "description": "Food DTO",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "foodName": {
                    "type": "string"
                },
                "foodType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "format": "double"
                },
                "archived": {
                    "type": "boolean"
                }
            }
        },
        "DailyMenu": {
            "type": "object",
            "description": "A daily Menu,  has order checked DTO",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "date": {
                    "type": "string",
                    "format": "date"
                },
                "orderId": {
                    "type": "integer",
                    "format": "int64"
                },
                "isFinal": {
                    "type": "boolean"
                },
                "foods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FoodWithQuantity"
                    }
                },
                "lastEdit": {
                    "$ref": "#/definitions/LastEdit"
                }
            }
        },
...
        "LastEdit": {
            "type": "object",
            "description": "Last edit DTO",
            "properties": {
                "timeStamp": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "number"
                }
            }
        },
...
    },
    "responses": {
        "Error": {
            "description": "An unexpected error occured.",
            "schema": {
                "$ref": "#/definitions/Error"
            }
        },
...
    }
}